[Feature] Architecture (#2)
