+++ client/src/components/Schedule/ScheduleCalendar.tsx
@@ -1,0 +1,3 @@
+export function ScheduleCalendar() {
+  return null;
+}
