+++ client/src/components/Schedule/ScheduleDetailView.tsx
@@ -1,0 +1,3 @@
+export function ScheduleDetailView() {
+  return null;
+}
