+++ client/src/components/Schedule/ScheduleList.tsx
@@ -1,0 +1,3 @@
+export function ScheduleList() {
+  return null;
+}
