+++ client/src/components/Schedule/ScheduleMap.tsx
@@ -1,0 +1,3 @@
+export function ScheduleMap() {
+  return null;
+}
