+++ client/src/components/Schedule/index.tsx
@@ -1,0 +1,6 @@
+//TODO: add lazy loading through @loadable
+export { ScheduleCalendar } from './ScheduleCalendar';
+export { ScheduleDetailView } from './ScheduleDetailView';
+export { ScheduleList } from './ScheduleList';
+export { ScheduleMap } from './ScheduleMap';
+export { ScheduleTable } from './ScheduleTable';
