+++ package.json
@@ -20,7 +20,6 @@
   "husky": {
     "hooks": {
       "pre-commit": "npm run format:check && npm run lint:js",
-      "pre-push": "npm run test",
       "post-commit": "git update-index --again"
     }
   },
